구조 변경 ** 기능 구현 - ** 수정 - SwaggerConfig.java : swagger 레퍼지토리 -> config 레퍼지토리로 moved ** 오류 해결 - ** 미해결 오류
