+++ src/main/java/jh/ParkingService/aws/S3Uploader.java
@@ -62,6 +62,35 @@ public class S3Uploader {
         return fileUrlList;
     }
 
+//    public List<String> updateFile(List<MultipartFile> multipartFiles, String dirName, String reviewerUid) {
+//        List<String> fileUrlList = new ArrayList<>(Arrays.asList(null,null,null,null,null));
+//
+//        //지정한 AWS S3 폴더에 있는 파일 초기화
+//        for(int i = 0;i< 5;i++){
+//            deleteFile(dirName + "/" + reviewerUid + "-" + i + ".jpg");
+//        }
+//
+//        // forEach 구문을 통해 multipartFile로 넘어온 파일들 하나씩 S3에 업로드하고 fileUrlList에 추가하여 리턴
+//        // MultipartFile 을 따로 File로 만드는 것이 아닌 InputStream 을 받는 방식
+//        multipartFiles.forEach(file -> {
+//            String fileName = createFileName(file.getOriginalFilename(), dirName, reviewerUid, cnt);
+//            ObjectMetadata objectMetadata = new ObjectMetadata();
+//            objectMetadata.setContentLength(file.getSize());
+//            objectMetadata.setContentType(file.getContentType());
+//
+//            try(InputStream inputStream = file.getInputStream()) {
+//                amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
+//                        .withCannedAcl(CannedAccessControlList.PublicRead));    //외부에 공개할 이미지이므로, 해당 파일에 public read 권한을 추가합니다.
+//            } catch(IOException e) {
+//                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
+//            }
+//            fileUrlList.add(cnt,amazonS3Client.getUrl(bucket, fileName).toString());
+//            cnt++;
+//        });
+//        cnt = 0;
+//        return fileUrlList;
+//    }
+
 
     public void deleteFile(String fileName) {
         amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
