+++ src/main/java/jh/ParkingService/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package jh.ParkingService.swagger;
+package jh.ParkingService.config;
 
 import java.util.ArrayList;
 import java.util.List;
